refactor(settings): extract the custom theme editor
Split the add-theme panel from the catalog page so color fields and save
actions live behind their own construction helpers.
